requirements.txt
googletrans
pillow
easyocr
deep-translator
black
opencv-python
librosa
